backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bitarray==3.8.1
certifi==2026.4.22
charset-normalizer==3.4.7
ckzg==2.1.7
click==8.3.3
colorama==0.4.6
cytoolz==1.1.0
eth-account==0.13.7
eth-hash==0.8.0
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==6.0.0
eth-utils==6.0.0
eth_abi==5.2.0
fastapi==0.136.1
frozenlist==1.8.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.13
multidict==6.7.1
parsimonious==0.10.0
propcache==0.4.1
pycryptodome==3.23.0
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
pyunormalize==17.0.0
pywin32==311
regex==2026.4.4
requests==2.33.1
rlp==4.1.0
starlette==1.0.0
toolz==1.1.0
types-requests==2.33.0.20260408
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.46.0
web3==7.15.0
websockets==15.0.1
yarl==1.23.0
